backend/requirements.txt
tensorflow
numpy
matplotlib
fastapi
uvicorn
pillow
python-multipart
opencv-python
tensorflow_datasets